fixup! fix(silo): correctly error when requested insertion positions are larger than reference sequence size
use uint32_t for cardinality to avoid exact overflow

+++ src/silo/storage/column/vertical_sequence_index.h
@@ -61,7 +61,7 @@ class VerticalSequenceIndex {
    static_assert(sizeof(SequenceDiffKey) == 8);
    struct SequenceDiff {
       roaring::internal::container_t* container;
-      uint16_t cardinality;
+      uint32_t cardinality;
       uint8_t typecode;
 
       template <class Archive>
